Interactive spectral map explorer

Shiny app to interactively explore spatially resolved hyperSpec objects.

spcmap_explorer(
  hyperspec_obj,
  fixed_y = TRUE,
  flip = FALSE,
  startband = 1600,
  metavar = NULL
)

Arguments

hyperspec_obj
A hyperSpec object with variables "x" and "y" in @data.

fixed_y
logical or numeric; if TRUE the initial y axis range is set to the intensity range of the complete dataset until the axis is manually modified or autoscale activated; alternatively provide a numeric vector of length 2 to set a fixed y range; if FALSE autoscale is activated.

flip
logical; if TRUE the x and y axes are transposed.

startband
numeric; band for the displayed image at startup.

metavar
character; a variable in @data can be selected as an alternative to the intensity map (e.g. to display cluster results).